WIP - re-working the file structure
WIP = Does Not Work right now

I misunderstood some pieces of how the gob encoding worked, and when I started
working on the file extraction hit issues. So after a bit of reading decided
the best course of action was to simplify. Now the file index at the end of the
archive contains all of the information about the files in the archive.

Right now the internal offset information in the index is not right, and I'm
not sure why. But I've been barred from any more technical work today so I can
have a hope at a good night's sleep.