Proper index storage after the files
The index is now at the end of the file, and the archive contains all of the
file contents. The checksum is not being stored properly, and I believe this
is a product of the []byte in the struct which is not being allocated properly
to be returned back to the caller.